Add mutex_lock model that can check multiple locks
The kernel-rules submodule now has 32_7 model for multiple-locking.  The
additional attributes are added to its "hints" section.  These attrs are
processed by blast backend and are transformed into new blast options.

Into the BLAST submodule were added capabilities to process these
options.  This functionality is commonly referred to as "Rerouter" or
"rerouting".

However, this is a heuristical approach, and the model is still not
precise.