Make preprocessing errors not crash all the tools
Preprocessing errors used to crash the whole stack with an integration
error.  Instead, it would be nice to just include these errors in the
final report, and just move on. This commit implements this in a
somewhat hacky manner.

We treat in this fix preprocessing error as an RCV error; however,
preprocessing was performed at DSCV stage (which is somewhat strange,
but DSCV/RCV distinction is a mess now).  When DSCV detects a
preprocessing error, it immediately sends a RESULT message for each of
its children (without issuing a relevant QUEUE command).  I fixed
cluster and local watcher backends to react to this as to a bit of a
usual workflow.  The results are then collected by DSCV and are
processed in a normal manner.

Also, I reorganized problems and added the CPP_error problem for
preprocessing errors.