Add CIL-based sample program analysis backend
It just demonstrates how you may plug in program analysis backends, and
run them on device drivers just as if a static analysis tool would have
been ran.

This particular backend collects all if statements with a pointer
equality test as a condition.  (Note that CIL unrolls shourt-circuit
evaluation to a series of if statements).

You may base your own measurements on this skeleton.  It's a bit "hacky"
but useful.