Add first version of RCV wrapper interface
RCV wrappers may now be implemented in concise terms, without making the
user interfere with various intrinsics of the integration with LDV and
reporting the result.

Preprocessing has been moved to RCV.

Still: preprocessing errors are not handled; CIL preprocessing is not
implemented, ETV is not moved to RCV (as it should be), file transfer
functionality is broken (i.e. the current launch works only locally).