Patch unionfs-fuse for faster preprocessing
The monotonic option was added to unionfs.

Define a monotonic file system as one on which files are only removed.
Then it's useless to try to open a file that wasn't located before.  We
create whiteouts for such files.

LDV cluster utilizes this option to improve speed of C preprocessing
over the network.